hostage_negotiator/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==5.1.7
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.67.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-http-client==3.3.7
sendgrid==6.11.0
sniffio==1.3.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tqdm==4.67.1
typing_extensions==4.12.2
